Change RTLD enums to static classes with public const int members.

+++ Impl/Helper.Linux.cs
@@ -7,7 +7,7 @@ namespace JetBrains.Profiler.Api.Impl
   {
     public static bool IsLibCoreApiSoAlreadyLoaded()
     {
-      var handle = LibDlSo2.dlopen(LibCoreApiSo.LibraryName, (int) (RTLD.RTLD_GLOBAL | RTLD.RTLD_LAZY | RTLD.RTLD_NOLOAD));
+      var handle = LibDlSo2.dlopen(LibCoreApiSo.LibraryName, RTLD.RTLD_GLOBAL | RTLD.RTLD_LAZY | RTLD.RTLD_NOLOAD);
       if (handle == IntPtr.Zero)
         return false;
       LibDlSo2.dlclose(handle);
