+++ Impl/Helper.MacOsX.cs
@@ -7,7 +7,7 @@ namespace JetBrains.Profiler.Api.Impl
   {
     public static bool IsLibCoreApiDylibAlreadyLoaded()
     {
-      var handle = LibDyldDylib.dlopen(LibCoreApiDylib.LibraryName, (int) (RTLD.RTLD_GLOBAL | RTLD.RTLD_LAZY | RTLD.RTLD_NOLOAD));
+      var handle = LibDyldDylib.dlopen(LibCoreApiDylib.LibraryName, RTLD.RTLD_GLOBAL | RTLD.RTLD_LAZY | RTLD.RTLD_NOLOAD);
       if (handle == IntPtr.Zero)
         return false;
       LibDyldDylib.dlclose(handle);
